Forbid deleting reviewed/published drafts and updates
Reviewed and/or published drafts and updates should be kept for record
keeping and so should not be deletable. More importantly, publishing
drafts and updates should not be deletable since deleting them breaks
server state.

+++ console/src/main/kotlin/app/accrescent/parcelo/console/routes/Updates.kt
@@ -358,9 +358,11 @@ fun Route.deleteUpdateRoute() {
             // the associated app
             call.respond(HttpStatusCode.NotFound, ApiError.updateNotFound(updateId))
         } else {
-            // If the user is the update's creator, delete the update. Otherwise, inform them they
+            // If the user is the update's creator and the update is neither reviewed, nor
+            // publishing, nor published, then delete the update. Otherwise, inform them that they
             // don't have sufficient permissions to delete the update.
-            if (update.creatorId == userId) {
+            val publishingOrPublished = update.submitted && update.reviewerId == null
+            if (update.creatorId == userId && update.reviewId == null && !publishingOrPublished) {
                 storageService.deleteFile(update.fileId)
                 transaction { update.delete() }
                 call.respond(HttpStatusCode.NoContent)
